Move watcher methods to subcommand
